a first draft of witness collection during generation
There is a problematic lost of typing information due to existentials:
  | Apply : ('a -> 'b) left_witness * 'a right_witness -> 'b left_witness

Even when one inspects the left_witness, and it may be evident what
the type of 'a is, it is not possible to statically reconstruct this
information.

Dynamic type-indexing schemes as discussed to reuse Fix may let us
recover this information dynamically. One good reason to tackle the
mutability problem first.